refactor: separate EventLabel component
